renamed Misguessed Summary to a more generic name, and added Vigi support
