+++ TownOfUs/Modifiers/Game/AssassinModifier.cs
@@ -1,13 +1,10 @@
-﻿using System.Collections;
-using HarmonyLib;
+﻿using HarmonyLib;
 using MiraAPI.GameOptions;
 using MiraAPI.Modifiers;
 using MiraAPI.Networking;
 using MiraAPI.Roles;
 using MiraAPI.Utilities;
 using MiraAPI.Utilities.Assets;
-using Reactor.Networking.Attributes;
-using Reactor.Networking.Rpc;
 using Reactor.Utilities;
 using TownOfUs.Modifiers.Crewmate;
 using TownOfUs.Modifiers.HnsGame;
@@ -202,7 +199,7 @@ public class AssassinModifier : TouGameModifier, IWikiDiscoverable
 
             if (ClickHandler(victim) && victim == Player)
             {
-                RpcAssassinMisguessSummary(Player, player.PlayerId, LastGuessedItem);
+                DeathHandlerModifier.RpcSetMisguessSummary(Player, player.PlayerId, LastGuessedItem);
             }
         }
 
@@ -217,7 +214,7 @@ public class AssassinModifier : TouGameModifier, IWikiDiscoverable
 
             if (ClickHandler(victim) && victim == Player)
             {
-                RpcAssassinMisguessSummary(Player, player.PlayerId, LastGuessedItem);
+                DeathHandlerModifier.RpcSetMisguessSummary(Player, player.PlayerId, LastGuessedItem);
             }
         }
 
@@ -280,32 +277,6 @@ public class AssassinModifier : TouGameModifier, IWikiDiscoverable
 
             shapeMenu.Close();
             return true;
-        }
-    }
-
-    [MethodRpc((uint)TownOfUsRpc.AssassinMisguessSummary, LocalHandling = RpcLocalHandling.After)]
-    public static void RpcAssassinMisguessSummary(PlayerControl assassin, byte victimId, string roleText)
-    {
-        var name = GameData.Instance?.GetPlayerById(victimId)?.Object?.Data?.PlayerName ?? "?";
-        var summary = TouLocale.GetParsed("TouModifierAssassinMisguessSummary")
-            .Replace("<player>", name)
-            .Replace("<role>", roleText);
-
-        Coroutines.Start(CoSetExtended(assassin, summary));
-    }
-
-    private static IEnumerator CoSetExtended(PlayerControl player, string summary)
-    {
-        var timeout = 5f;
-        while (timeout > 0f && player && !player.HasModifier<DeathHandlerModifier>())
-        {
-            timeout -= Time.deltaTime;
-            yield return null;
-        }
-
-        if (player && player.TryGetModifier<DeathHandlerModifier>(out var deathHandler))
-        {
-            deathHandler.ExtendedCauseOfDeath = summary;
         }
     }
