+++ TownOfUs/Roles/Classic/Crewmate/CrewmateKilling/VigilanteRole.cs
@@ -8,6 +8,7 @@ using MiraAPI.Patches.Stubs;
 using MiraAPI.Roles;
 using MiraAPI.Utilities;
 using Reactor.Utilities;
+using TownOfUs.Modifiers;
 using TownOfUs.Modifiers.Crewmate;
 using TownOfUs.Modifiers.Game;
 using TownOfUs.Modifiers.HnsGame;
@@ -169,19 +170,28 @@ public sealed class VigilanteRole(IntPtr cppPtr) : CrewmateRole(cppPtr), ITouCre
                 }
             }
             var victim = pickVictim ? player : Player;
+            var roleText = $"{role.TeamColor.ToTextColor()}{role.GetRoleName()}</color>";
 
-            ClickHandler(victim);
+            if (ClickHandler(victim) && victim == Player)
+            {
+                DeathHandlerModifier.RpcSetMisguessSummary(Player, player.PlayerId, roleText);
+            }
         }
 
         void ClickModifierHandle(BaseModifier modifier)
         {
             var pickVictim = player.HasModifier(modifier.TypeId);
             var victim = pickVictim ? player : Player;
+            var modifierText =
+                $"{MiscUtils.GetRoleColour(modifier.ModifierName.Replace(" ", string.Empty)).ToTextColor()}{modifier.ModifierName}</color>";
 
-            ClickHandler(victim);
+            if (ClickHandler(victim) && victim == Player)
+            {
+                DeathHandlerModifier.RpcSetMisguessSummary(Player, player.PlayerId, modifierText);
+            }
         }
 
-        void ClickHandler(PlayerControl victim)
+        bool ClickHandler(PlayerControl victim)
         {
             if (!OptionGroupSingleton<VigilanteOptions>.Instance.VigilanteMultiKill || MaxKills == 0 ||
                 victim == Player)
@@ -197,7 +207,7 @@ public sealed class VigilanteRole(IntPtr cppPtr) : CrewmateRole(cppPtr), ITouCre
 
                 shapeMenu.Close();
 
-                return;
+                return false;
             }
 
             if (victim == Player && SafeShotsLeft != 0)
@@ -213,7 +223,7 @@ public sealed class VigilanteRole(IntPtr cppPtr) : CrewmateRole(cppPtr), ITouCre
 
                 shapeMenu.Close();
 
-                return;
+                return false;
             }
             Player.RpcSpecialMurder(victim, MeetingCheck.ForMeeting, true, true, createDeadBody: false, teleportMurderer: false,
                 showKillAnim: false,
@@ -226,6 +236,7 @@ public sealed class VigilanteRole(IntPtr cppPtr) : CrewmateRole(cppPtr), ITouCre
             }
 
             shapeMenu.Close();
+            return true;
         }
     }
